멤버 엔티티,dto 성별, 나이 추가 관리자페이지 고객 목록 추가 완료 (조건 검색 미완) 멤버 추가 테스트 수정 완.
4월 26
회원가입 html 유효성 검사 일부 완
admin,mypage 컨트롤러 분리
mypage 레이아웃 완
마이페이지 내 정보 가기위한 pwCheck 완

해야될것. 내 정보 수정
