+++ shop/src/main/java/com/keduit/shop/controller/MemberController.java
@@ -89,9 +89,14 @@ public class MemberController {
     }
 
     /*비밀번호 찾기*/
+//    @GetMapping("/findLoginPw")
+//    public String findLoginPw(){
+//        return "member/findLoginPw";
+//    }
+
     @GetMapping("/findLoginPw")
     public String findLoginPw(){
-        return "member/findLoginPw";
+        return "member/pwInquiry";
     }
 
 /*    @GetMapping("/doFindLoginPw")
@@ -100,29 +105,12 @@ public class MemberController {
             return Utility.jsHistoryBack("이메일을 입력해주세요");
         }
     }*/
-    @GetMapping({"/admin/{page}", "/admin"})
-    public String memberMangeListPage(Model model,
-                                      @PathVariable("page") Optional<Integer> page, // 유저에게 받는 page 숫자.
-                                      AdminMemberSearchDTO adminMemberSearchDTO) { //쿼리문을 날리기 위한 정보들!
 
+//    ------------------- 마이페이지 컨트롤러 ----------------------
 
-
-        Pageable pageable = PageRequest.of(page.isPresent() ? page.get() : 0, 10);
-
-
-        Page<Member> members = memberService.getAdminMemberPage(adminMemberSearchDTO, pageable);
-
-        System.out.println("----- members.getContent() : " + members.getContent());
-        System.out.println("----- adminMemberSearchDTO: " + adminMemberSearchDTO);
-
-        model.addAttribute("members", members);
-        model.addAttribute("adminMemberSearchDTO", adminMemberSearchDTO);
-        model.addAttribute("maxPage", 10);
-
-        System.out.println(members.getNumber()+"@@@@@@@@@@@@@@@@@@@@@@");
-
-
-        return "admin/memberMng";
+    @GetMapping("/checkPw")
+    public String checkPw(){
+        return "mypage/checkPw";
     }
 
 }
