+++ shop/src/main/java/com/keduit/shop/controller/mypage/MypageOrderController.java
@@ -1,0 +1,4 @@
+package com.keduit.shop.controller.mypage;
+
+public class MypageOrderController {
+}
