+++ shop/src/main/java/com/keduit/shop/service/MemberService.java
@@ -1,6 +1,7 @@
 package com.keduit.shop.service;
 
 import com.keduit.shop.dto.AdminMemberSearchDTO;
+import com.keduit.shop.dto.MemberFormDTO;
 import com.keduit.shop.entity.Member;
 import com.keduit.shop.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
@@ -10,6 +11,7 @@ import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 
@@ -20,6 +22,7 @@ import javax.transaction.Transactional;
 @RequiredArgsConstructor//필요한 부분만 생성자로
 public class MemberService implements UserDetailsService {
     private final MemberRepository memberRepository;
+    private final PasswordEncoder passwordEncoder;
 
     public Member saveMember(Member member) {//새로운 멤버를 등록할 시 중복x
         validateDuplicateMember(member);//중복 여부를 검증하는 과정
@@ -52,6 +55,13 @@ public class MemberService implements UserDetailsService {
 
     public Page<Member> getAdminMemberPage(AdminMemberSearchDTO adminMemberSearchDTO, Pageable pageable) {
         return memberRepository.getAdminMemberPage(adminMemberSearchDTO, pageable);
+    }
+
+    public boolean checkPw(String getPass,String getName) {
+        Member member = memberRepository.findByEmail(getName);
+
+        return passwordEncoder.matches(getPass,member.getPassword());
+
     }
 }
