+++ shop/src/main/resources/templates/fragments/adminSidebar.html
@@ -27,7 +27,7 @@
         </button>
         <div class="dropdown-container">
             <a href="managerCs.html">고객 문의</a>
-             <a href="/members/admin">고객 목록</a>
+             <a href="/admin/members">고객 목록</a>
         </div>
         <button class="dropdown-btn">판매자 정보
         </button>
