+++ shop/src/main/resources/templates/item/itemDtl.html
@@ -73,7 +73,12 @@
                             <dt>판매가</dt>
                             <dd class="sale">
 
+
+
                                 <em>208,050</em> 원 <p class="sale_per">5%</p>
+
+
+
                             </dd>
                         </dl>
                     </div>
