+++ shop/src/main/resources/templates/mypage/checkPw.html
@@ -1,0 +1,28 @@
+<!DOCTYPE html>
+<html lang="ko" xmlns:th="http://www.thymeleaf.org"
+      xmlns:layout=http://www.ultraq.net.nz/thymeleaf/layout
+      layout:decorate='~{layouts/mypageLayout}'>
+
+<div layout:fragment="content">
+    <link th:href="@{/css/mypage/mypageCheckPw.css}" rel="stylesheet" type="text/css">
+
+    <form th:action="@{/mypage/checkPw}" role="form" method="post" th:object="${memberFormDTO}">
+
+    <div class="mainSection2">
+        <h3>비밀번호 확인</h3>
+        <div class="pwdCheckBox">
+            <p>회원님의 안전한 개인정보 보호를 위해 비밀번호를 다시 한번 확인합니다. </p>
+            <div class="inpBox">
+                <input type="password" th:name="password" placeholder="비밀번호를 입력해주세요">
+                <a href="myPageChange.html"><button>확인</button></a>
+            </div>
+<!--            <div th:if="${errorMessage}" class="alert alert-danger" role="alert">-->
+<!--                <p th:text="${errorMessage}"></p>-->
+<!--            </div>-->
+        </div>
+    </div>
+    <div class="mainSection1"></div>
+
+    </form>
+</div>
+</html>
